调整监控机制，并修改样例。

+++ ps-core/src/main/java/org/ps/platform/core/Task.java
@@ -73,7 +73,7 @@ public class Task implements Serializable, Comparable<Task> {
     /**
      * 进度信息
      */
-    private int process;
+    private float process;
 
     /**
      * 是否暂停
